refactor: Add playlistReducer to reducer folder
